pyproject.toml
[project]
name="replit-river"
version="0.0.0a0"
description="Replit river toolkit for Python"
authors = [{ name = "Replit", email = "eng@replit.com" }]
license = { file = "LICENSE" }
keywords = ["rpc", "websockets"]
readme = "README.md"
requires-python = ">=3.11,<4"
classifiers = [
  "Development Status :: 4 - Beta",

  "Intended Audience :: Developers",
  "Topic :: Internet",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic==2.9.2",
    "aiochannel>=1.2.1",
    "grpcio-tools>=1.59.3",
    "grpcio>=1.59.3",
    "msgpack>=1.0.7",
    "nanoid>=2.0.0",
    "protobuf>=5.28.3",
    "pydantic-core>=2.20.1",
    "websockets>=12.0",
    "opentelemetry-sdk>=1.28.2",
    "opentelemetry-api>=1.28.2",
]

[tool.uv]
dev-dependencies = [
    "deptry>=0.14.0",
    "msgpack-types>=0.3.0",
    "mypy>=1.4.0",
    "mypy-protobuf>=3.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.11.1",
    "ruff>=0.0.278",
    "types-protobuf>=4.24.0.20240311",
    "types-nanoid>=2.0.0.20240601",
    "pyright>=1.1.389",
    "pytest-snapshot>=0.9.0",
]

[tool.ruff]
lint.select = ["F", "E", "W", "I001"]
exclude = ["*/generated/*", "*/snapshots/*"]

# Should be kept in sync with mypy.ini in the project root.
# The VSCode mypy extension can only read /mypy.ini.
# While mypy run inside the chat container can only see this file.
[tool.mypy]
plugins = "pydantic.mypy"
disallow_untyped_defs = true
warn_return_any = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"  # auto-detect async tests/fixtures
addopts = "--tb=short"
env = [
	"DD_DOGSTATSD_DISABLE=true",
	"DD_TRACE_ENABLED=false",
]
filterwarnings = [
	"ignore::DeprecationWarning",  # google SDKs cause this noise
]
markers = [
	"e2e: marks tests as end-to-end (deselect with '-m \"not e2e\"')",
]

[[tool.mypy.overrides]]
module = [
	"google.auth.*",
	"google.oauth2.*",
	"google.cloud.sqlalchemy_spanner.sqlalchemy_spanner.*",
	"grpc.*",
	"grpc_tools.*",
	"nanoid.*",
]
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/replit_river"]
